+++ lilfilter/resampler.py
@@ -195,7 +195,7 @@ class Resampler:
             # will be shape (minibatch_size, in_channels, seq_len) with in_channels == 1
             in_data = in_data.unsqueeze(1)
             out = torch.nn.functional.conv1d(in_data,
-                                             self.weights,
+                                             self.weights.type_as(in_data),
                                              stride=self.input_sr,
                                              padding=self.padding)
             # shape will be (minibatch_size, out_channels = 1, seq_len);
@@ -203,7 +203,7 @@ class Resampler:
             return out.squeeze(1)
         elif self.resample_type == 'integer_upsample':
             out = torch.nn.functional.conv_transpose1d(in_data.unsqueeze(1),
-                                                       self.weights,
+                                                       self.weights.type_as(in_data),
                                                        stride=self.output_sr,
                                                        padding=self.padding)
             return out.squeeze(1)
@@ -221,7 +221,7 @@ class Resampler:
             # in_channels, width) so we need to reshape (note: time is width).
             in_data = in_data.transpose(1, 2)
 
-            out = torch.nn.functional.conv1d(in_data, self.weights,
+            out = torch.nn.functional.conv1d(in_data, self.weights.type_as(in_data),
                                              padding=self.padding)
             assert out.shape == (minibatch_size, self.output_sr, num_blocks)
             return out.transpose(1, 2).contiguous().view(minibatch_size, num_blocks * self.output_sr)
